Avoid saying "successful return" when describing a non-throwing Start()
AsyncJob::Start() return value is not yet documented. When it is
documented, it is likely to promise to never return falsy results,
similar to an always-successful new expression. Otherwise, the callers
following the startStepN() sketch will wait for a callback that may
never come. That infinite wait is one of the biggest problems the
documented principles are trying to avoid.